Fixes test URLs and dimensions
The tests in this repo are _bad_. Like, really bad.

All of the request for external assets should be mocked, probably
by setting up a small HTTP server to test against. This was only
put together as an experiment, so I can't say I was hugely bothered
at the time. I'd fix this if I was to get this app ready for production
though.

PRs to fix are welcome.